refactor: resolve ESLint errors, enhance type safety, and decouple routes for Fast Refresh
