+++ eslint.config.js
@@ -6,7 +6,7 @@ import tseslint from 'typescript-eslint'
 import { defineConfig, globalIgnores } from 'eslint/config'
 
 export default defineConfig([
-  globalIgnores(['dist']),
+  globalIgnores(['dist', 'coverage', '.output.txt', 'eslint_errors.txt']),
   {
     files: ['**/*.{ts,tsx}'],
     extends: [
