+++ package.json
@@ -7,7 +7,7 @@
     "dev": "vite",
     "build": "vite build && tsc -b",
     "preview": "vite preview",
-    "lint": "biome check .",
+    "lint": "eslint . && biome check .",
     "format": "biome format --write .",
     "test": "vitest run",
     "test:watch": "vitest",
