+++ src/components/layout/MobileNav.tsx
@@ -1,7 +1,7 @@
 import { Link } from '@tanstack/react-router'
 import { LayoutDashboard, Tag, ArrowLeftRight, Settings } from 'lucide-react'
 import { cn } from '@/lib/cn'
-import { useState } from 'react'
+import { useCallback, useRef } from 'react'
 
 const NAV_ITEMS = [
   { to: '/', label: 'Dashboard', icon: LayoutDashboard },
@@ -11,16 +11,15 @@ const NAV_ITEMS = [
 ] as const
 
 export function MobileNav() {
-  const [lastTap, setLastTap] = useState<{ [key: string]: number }>({})
+  const lastTapRef = useRef<{ [key: string]: number }>({})
 
-  const handleTap = (to: string) => {
-    const now = Date.now()
-    const last = lastTap[to] || 0
-    if (now - last < 300) {
+  const handleTap = useCallback((to: string, timeStamp: number) => {
+    const last = lastTapRef.current[to] || 0
+    if (timeStamp - last < 300) {
       window.scrollTo({ top: 0, behavior: 'smooth' })
     }
-    setLastTap((prev) => ({ ...prev, [to]: now }))
-  }
+    lastTapRef.current[to] = timeStamp
+  }, [])
 
   return (
     <nav className="fixed bottom-0 left-0 right-0 z-50 border-t bg-background pb-[env(safe-area-inset-bottom)] md:hidden">
@@ -32,7 +31,7 @@ export function MobileNav() {
             className="flex flex-1 flex-col items-center gap-0.5 py-1.5 text-xs text-muted-foreground transition-colors hover:text-foreground"
             activeProps={{ className: 'text-primary font-semibold' }}
             activeOptions={{ exact: to === '/' }}
-            onClick={() => handleTap(to)}
+            onClick={(e) => handleTap(to, e.timeStamp)}
           >
             <Icon className="h-5 w-5" />
             <span className="text-xs">{label}</span>
