+++ src/components/layout/RootComponent.tsx
@@ -1,0 +1,21 @@
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
+import { Outlet } from '@tanstack/react-router'
+import { TanStackRouterDevtools } from '@tanstack/react-router-devtools'
+import { Toaster } from '@/components/ui/toaster'
+import { VersionCheck } from '@/components/VersionCheck'
+
+export function RootComponent() {
+  return (
+    <>
+      <VersionCheck />
+      <Outlet />
+      <Toaster />
+      {import.meta.env.DEV && (
+        <>
+          <TanStackRouterDevtools position="bottom-right" />
+          <ReactQueryDevtools initialIsOpen={false} />
+        </>
+      )}
+    </>
+  )
+}
