+++ src/components/transactions/TransactionList.test.tsx
@@ -1,6 +1,7 @@
 import { render, screen, fireEvent } from '@testing-library/react'
 import { TransactionList } from './TransactionList'
 import { describe, expect, it, vi } from 'vitest'
+import type { Transaction } from '@budget-buddy-org/budget-buddy-contracts'
 
 vi.mock('@/hooks/use-toast', () => ({
   useToast: () => ({ toast: vi.fn() }),
@@ -34,11 +35,11 @@ describe('TransactionList empty states', () => {
   })
 
   it('groups transactions by date', () => {
-    const transactions = [
+    const transactions: Transaction[] = [
       { id: '1', description: 'Item 1', date: '2024-01-01', amount: 1000, currency: 'EUR', type: 'EXPENSE', categoryId: 'c1' },
       { id: '2', description: 'Item 2', date: '2024-01-01', amount: 2000, currency: 'EUR', type: 'INCOME', categoryId: 'c1' },
       { id: '3', description: 'Item 3', date: '2024-01-02', amount: 3000, currency: 'EUR', type: 'EXPENSE', categoryId: 'c1' },
-    ] as any[]
+    ]
     const categories = [{ id: 'c1', name: 'Category 1' }]
     
     render(<TransactionList transactions={transactions} categories={categories} isLoading={false} />)
