+++ src/components/ui/badge.tsx
@@ -1,30 +1,9 @@
-import { type VariantProps, cva } from 'class-variance-authority'
 import * as React from 'react'
 import { cn } from '@/lib/cn'
-
-const badgeVariants = cva(
-  'inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors',
-  {
-    variants: {
-      variant: {
-        default: 'border-transparent bg-primary text-primary-foreground',
-        secondary: 'border-transparent bg-secondary text-secondary-foreground',
-        destructive: 'border-transparent bg-destructive text-destructive-foreground',
-        outline: 'text-foreground',
-        income: 'border-transparent bg-income/15 text-income',
-        expense: 'border-transparent bg-muted text-foreground',
-      },
-    },
-    defaultVariants: { variant: 'default' },
-  },
-)
-
-export interface BadgeProps
-  extends React.HTMLAttributes<HTMLDivElement>,
-    VariantProps<typeof badgeVariants> {}
+import { badgeVariants, type BadgeProps } from './badge-variants'
 
 function Badge({ className, variant, ...props }: BadgeProps) {
   return <div className={cn(badgeVariants({ variant }), className)} {...props} />
 }
 
-export { Badge, badgeVariants }
+export { Badge }
