+++ src/components/ui/date-picker.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { Input } from './input'
 import { cn } from '@/lib/cn'
 
-export interface DatePickerProps extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'type'> {}
+export type DatePickerProps = Omit<React.InputHTMLAttributes<HTMLInputElement>, 'type'>
 
 const DatePicker = React.forwardRef<HTMLInputElement, DatePickerProps>(
   ({ className, ...props }, ref) => {
