+++ src/routes/_app/transactions/index.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/_app/transactions/')({
-  validateSearch: (search: Record<string, unknown>) => ({
+  validateSearch: () => ({
     // Previous parameters were add/edit, now they are no longer used for routing to the modal
   }),
 })
