+++ src/routes/_auth.tsx
@@ -1,5 +1,6 @@
-import { Outlet, createFileRoute, redirect } from '@tanstack/react-router'
+import { createFileRoute, redirect } from '@tanstack/react-router'
 import { useAuthStore } from '@/stores/auth.store'
+import { AuthLayout } from '@/components/layout/AuthLayout'
 
 export const Route = createFileRoute('/_auth')({
   beforeLoad: () => {
@@ -9,17 +10,3 @@ export const Route = createFileRoute('/_auth')({
   },
   component: AuthLayout,
 })
-
-function AuthLayout() {
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-background p-4">
-      <div className="w-full max-w-sm">
-        <div className="mb-8 text-center">
-          <h1 className="text-2xl font-semibold tracking-tight">Budget Buddy</h1>
-          <p className="mt-1 text-sm text-muted-foreground">Your personal finance companion</p>
-        </div>
-        <Outlet />
-      </div>
-    </div>
-  )
-}
