+++ src/routes/_auth/login.a11y.test.tsx
@@ -2,12 +2,12 @@ import { render } from '@testing-library/react'
 import 'vitest-axe/extend-expect'
 import { axe } from 'vitest-axe'
 import { describe, expect, it, vi } from 'vitest'
-import { Route } from './login.lazy'
+import { LoginPage } from '@/components/auth/LoginPage'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import type React from 'react'
+import React from 'react'
 
 vi.mock('@tanstack/react-router', () => ({
-  createLazyFileRoute: () => (options: any) => ({ options }),
+  createLazyFileRoute: () => (options: { component: React.ComponentType }) => ({ options }),
   useNavigate: () => vi.fn(),
   Link: ({ children }: { children: React.ReactNode }) => <a>{children}</a>,
 }))
@@ -21,8 +21,6 @@ describe('LoginPage a11y', () => {
         },
       },
     })
-    
-    const LoginPage = (Route as any).options.component as React.ElementType
     
     const { container } = render(
       <QueryClientProvider client={queryClient}>
