+++ src/routes/_auth/register.a11y.test.tsx
@@ -2,12 +2,12 @@ import { render } from '@testing-library/react'
 import 'vitest-axe/extend-expect'
 import { axe } from 'vitest-axe'
 import { describe, expect, it, vi } from 'vitest'
-import { Route } from './register.lazy'
+import { RegisterPage } from '@/components/auth/RegisterPage'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import type React from 'react'
+import React from 'react'
 
 vi.mock('@tanstack/react-router', () => ({
-  createLazyFileRoute: () => (options: any) => ({ options }),
+  createLazyFileRoute: () => (options: { component: React.ComponentType }) => ({ options }),
   useNavigate: () => vi.fn(),
   Link: ({ children }: { children: React.ReactNode }) => <a>{children}</a>,
 }))
@@ -21,8 +21,6 @@ describe('RegisterPage a11y', () => {
         },
       },
     })
-    
-    const RegisterPage = (Route as any).options.component as React.ElementType
     
     const { container } = render(
       <QueryClientProvider client={queryClient}>
